Restructure country pages
- Separate out routes into multiple files
- Add dummy views for new country pages
