+++ src/components/domain/CountryPageEmptyMessage/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "common",
+    "strings": {
+        "countryPageWipMessage": "This page is currently under construction!"
+    }
+}
