+++ src/components/domain/CountryPageEmptyMessage/styles.module.css
@@ -1,0 +1,3 @@
+.country-page-empty-message {
+    min-height: 24rem;
+}
