+++ src/views/CountryNsOverview/styles.module.css
@@ -1,0 +1,3 @@
+.country-ns-overview {
+    padding: var(--go-ui-spacing-2xl) 0;
+}
