+++ src/views/CountryNsOverviewActivities/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryNsOverviewActivities",
+    "strings": {
+        "pageTitle": "Country NS Activities"
+    }
+}
