+++ src/views/CountryNsOverviewCapacity/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryNsOverviewCapacity",
+    "strings": {
+        "pageTitle": "Country NS Capacity"
+    }
+}
