+++ src/views/CountryNsOverviewContextAndStructure/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryNsOverviewContextAndStructure",
+    "strings": {
+        "pageTitle": "Country NS Context and Structure"
+    }
+}
