+++ src/views/CountryNsOverviewStrategicPriorities/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryNsOverviewStrategicPriorities",
+    "strings": {
+        "pageTitle": "Country NS Strategic Priorities"
+    }
+}
