+++ src/views/CountryOngoingActivities/styles.module.css
@@ -1,0 +1,3 @@
+.country-ongoing-activities {
+    padding: var(--go-ui-spacing-2xl) 0;
+}
