+++ src/views/CountryOngoingActivitiesEmergencies/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryOngoingActivitiesEmergencies",
+    "strings": {
+        "pageTitle": "Country Ongoing Emergencies"
+    }
+}
