+++ src/views/CountryOngoingActivitiesThreeWActivities/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryOngoingActivitiesThreeWActivities",
+    "strings": {
+        "pageTitle": "Country Ongoing 3W Activities"
+    }
+}
