+++ src/views/CountryOngoingActivitiesThreeWProjects/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryOngoingActivitiesThreeWProjects",
+    "strings": {
+        "pageTitle": "Country 3W Projects"
+    }
+}
