+++ src/views/CountryProfile/i18n.json
@@ -1,0 +1,9 @@
+{
+    "namespace": "countryProfile",
+    "strings": {
+        "overviewTabTitle": "Overview",
+        "supportingPartnersTabTitle": "Supporting Partners",
+        "previousEventsTabTitle": "Previous Events",
+        "seasonalRisksTabTitle": "Seasonal Risks"
+    }
+}
