+++ src/views/CountryProfile/styles.module.css
@@ -1,0 +1,3 @@
+.country-profile {
+    padding: var(--go-ui-spacing-2xl) 0;
+}
