+++ src/views/CountryProfileOverview/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryProfileOverview",
+    "strings": {
+        "pageTitle": "Country Profile Overview"
+    }
+}
