+++ src/views/CountryProfilePreviousEvents/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryProfilePreviousEvents",
+    "strings": {
+        "pageTitle": "Country Profile Previous Events"
+    }
+}
