+++ src/views/CountryProfileSeasonalRisks/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryProfileSeasonalRisks",
+    "strings": {
+        "pageTitle": "Country Profile Seasonal Risks"
+    }
+}
