+++ src/views/CountryProfileSupportingPartners/i18n.json
@@ -1,0 +1,6 @@
+{
+    "namespace": "countryProfileSupportingPartners",
+    "strings": {
+        "pageTitle": "Country Profile Supporting Partners"
+    }
+}
